More progress.

+++ func_tbl.h
@@ -2,9 +2,9 @@
 #define func_tbl_H
 
 extern const u8 far functbl[];
-extern const u8 far tuneup_table[];
+extern const u16 far tuneup_table[];
 extern const char far * far weapon_name[];
-extern const u8 far aulist[];
+extern const s8 far aulist[];
 extern const char far * far chara_name[];
 extern const char far * far type_name[];
 extern const u8 far stagelist[];
